% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.grepData}
\alias{.grepData}
\title{Helper function for parsing output of samtools stats}
\usage{
.grepData(input, section, columns, col.names = "")
}
\arguments{
\item{input}{A character vector containing output of samtools stats, each line as one element.}

\item{section}{A pattern to extract from the report. Each section of samtools stats report is uniquley
labeled, for example Summary Numbers are labeled as SN. Currently supported sections: SN, IS.}

\item{columns}{A integer vector containing numbers of columns to extract from specified section of samtools
stats report.}

\item{col.names}{A character vector containing names of extracted columns. Optional.}
}
\value{
data frame holding data from selected section of \code{samtools stat} output.
}
\description{
As readSamtoolsStats parses output of samtools stat this function is used to grep relevant
parts of the output.
}
\examples{
file <- "path/to/samtools/stats/output"
inputFile <- readLines(file)
.grepData(input=inputFile, section="SN", columns=c(2,3), col.names=c("description", "value"))

}
