% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotInsertSize}
\alias{plotInsertSize}
\title{Plot Insert Size Samtools Stat section}
\usage{
plotInsertSize(data, samples, log = FALSE, lims = c(0, 400),
  sizeLimit = 75)
}
\arguments{
\item{data}{Insert Size (IS)) part of output returned by \code{\link{readSamtoolsStats}}
or formated data frame of values to plot.}

\item{samples}{vector of samples names from data that are required to be plotted.}

\item{log}{logical value indicating if data should be log transformed before plotting.}

\item{lims}{two element vector specifing x-axis limists.}

\item{sizeLimit}{lower limit on insert size. Data below this limit will be filtered out.}
}
\description{
plotInsertSize plots Insert Size (IS) part of output returned by
\code{\link{readSamtoolsStats}}.
}
\examples{
IS <- readSamtoolsStats()
plotInsertSize <- function(data, samples, log = FALSE, lims = c(0, 400), sizeLimit = 75)

}
