% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMixtureDistribution}
\alias{plotMixtureDistribution}
\title{Plot mixture of exponent and gaussian distributions}
\usage{
plotMixtureDistribution(x, start, gnumber = 3, sizeLimit = 75)
}
\arguments{
\item{x}{numeric vector of insert size}

\item{start}{list of parameters for exponential and gaussian functions. See example.}

\item{gnumber}{number of gaussian functions to use.}

\item{sizeLimit}{lower limit on insert size. Data below this limit will be filtered out.}
}
\description{
Plots mixture of exponent and gaussian distributions that apporximate the insert size distribution
in ATAC-seq experiment.
}
\examples{
x <- 75:1000
plotMixtureDistribution(x, start = list(a = 0.1, b = -0.1, k1 = 2e-5, mean1 = 200, sd1 = 15, k2 = 2e-5, mean2 = 400, sd2 = 15), gnumber = 2)

}
