% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{readSamtoolsStats}
\alias{readSamtoolsStats}
\title{Parse samtools stat output}
\usage{
readSamtoolsStats(file, ...)
}
\arguments{
\item{file}{the name of the file which the data are to be read from. File must be
of type \code{character}, providing path to input file. If it does not provide an
absolute path(s), the file name is relative to the current working directory, \code{getwd()}.}
}
\value{
list of data frames holding data from different parts of \code{samtools stat} output.
}
\description{
readSamtoolsStats parses output of samtools stat making it easy to
work with it in R.
}
\examples{
readSamtoolsStats(file = "path/to/samtools/stats/output")

}
