% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACapp.R
\name{runATACapp}
\alias{runATACapp}
\title{Run ATAC-seq quality control application}
\usage{
runATACapp()
}
\description{
Launch Shiny application that allows to quickly insepect multiple ATAC-seq
libraries and check some of the basic quality controls.
}
\examples{
sumsamstats::runATACapp()

}
