% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsamstats.R
\docType{package}
\name{sumsamstats}
\alias{sumsamstats}
\alias{sumsamstats-package}
\title{sumsamstats: A package for parsing and summarizing output of samtools stat,
with support for multiple files.}
\description{
The \code{sumsamstats} package provides parser that allows to parse output of
\code{samtools stat} and multiple plotting functions which make it easy to
plot parsed statistics for quick quality inspection. Package is shipped with
aplication that allows checking multiple quality parameters of ATAC-seq libraries.
}
