% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSummaryNumbers}
\alias{plotSummaryNumbers}
\title{Plot Summary Numbers Samtools Stat section}
\usage{
plotSummaryNumbers(data, samples, what = "raw total sequences")
}
\arguments{
\item{data}{Summary Numbers (SN) part of output returned by \code{\link{readSamtoolsStats}}
or formated data frame of values to plot.}

\item{samples}{vector of samples names from data that are required to be plotted.}

\item{what}{string describing property to plot from Summary Numbers.}
}
\description{
plotSummaryNumbers plots Summary Numbers (SN) part of output returned by
\code{\link{readSamtoolsStats}}.
}
\examples{
SN <- readSamtoolsStats()
plotSummaryNumbers(data = SN, samples = c("sample1", "sample2"), what = "raw total sequences")

}
