% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{readSamtoolsStats}
\alias{readSamtoolsStats}
\title{Parse samtools stat output}
\usage{
readSamtoolsStats(file, ...)
}
\arguments{
\item{file}{the name of the file which the data are to be read from. File must be
of type \code{character} and can be a vector of characters, providing paths to
multiple input files. If it does not provide an absolute path(s), the file name is
relative to the current working directory, \code{getwd()}.}
}
\value{
If one file is provided as input returns list of data frames holding data
from different parts of \code{samtools stat} output. In case that multiple files are
provided list of lists is returned.
}
\description{
readSamtoolsStats parses output of samtools stat making it easy to
work with it in R. Supports reading multiple files at once.
}
\examples{
readSamtoolsStats(file = "path/to/samtools/stats/output")

}
