% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{readSamtoolsStats}
\alias{readSamtoolsStats}
\title{Parse samtools stat output}
\usage{
readSamtoolsStats(file, section = c("SN", "FFQ", "LFQ", "GCF", "GCL",
  "GCC", "IS", "RL", "ID", "IC", "COV", "GCD"))
}
\arguments{
\item{file}{the name of the file which the data are to be read from. File must be
of type \code{character}, providing path to input file. If it does not provide an
absolute path(s), the file name is relative to the current working directory, \code{getwd()}.}

\item{section}{A character vector of patterns to extract from the report. Each section of samtools stats report is uniquley
labeled, for example Summary Numbers are labeled as SN. By default all sections are read.}
}
\value{
list of data frames holding data from different parts of \code{samtools stat} output.
Sections names are: SN (summary numbers), FFQ (first fragment qualities), LFQ (last fragment qualities),
GCF (GC Content of first fragments), GCL (GC content of last fragments), GCC (ACGT content per cycle),
IS (insert size), RL (read lengths), ID (indel distribution), IC (indels per cycle), COV (coverage distribution),
GCD (GC-depth)
}
\description{
readSamtoolsStats parses output of samtools stat making it easy to
work with it in R.
}
\examples{
readSamtoolsStats(file = system.file("extdata", "sample1", package = "sumsamstats"))

}
